requirements.txt
alembic==1.16.5
bibtexparser==1.4.3
blinker==1.9.0
click==8.3.0
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
Pillow==10.0.0
pyparsing==3.2.5
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
Werkzeug==2.3.6
WTForms==3.2.1
